Fix up cursor bounds checking
	Now when we move the cursor around, it is bounded on a line by
line basis rather than by the screen dimensions. Also if you press left
on the first character in a line, it wraps around to the last character
in the previous line. And if you press right on the last character in a
line, it wraps around to the first character in the next line.